Okay, that's working. Good work ! <3

+++ GameEngine/Vendors/TinyObjLoader/tiny_obj_loader.cpp
@@ -168,7 +168,7 @@ updateVertex(
     return it->second;
   }
 
-  assert(in_positions.size() > (3*i.v_idx+2));
+  assert(in_positions.size() > size_t(3*i.v_idx+2));
 
   positions.push_back(in_positions[3*i.v_idx+0]);
   positions.push_back(in_positions[3*i.v_idx+1]);
